[Rename : back] GoogleConfigUtil -> GoogleConfig로 변경
