+++ backend/spring-boot-3/src/main/java/back/springbootdeveloper/seungchan/config/GoogleConfig.java
@@ -1,23 +1,13 @@
-package back.springbootdeveloper.seungchan.util;
+package back.springbootdeveloper.seungchan.config;
 
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
-import org.springframework.web.util.DefaultUriBuilderFactory;
 import org.springframework.web.util.UriComponentsBuilder;
-
-import java.net.URI;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 @Component
 @Data
-public class GoogleConfigUtil {
+public class GoogleConfig {
     @Value("${google.auth.url}")
     private String googleAuthUrl;
 
@@ -35,5 +25,13 @@ public class GoogleConfigUtil {
 
     @Value("${google.auth.scope}")
     private String scopes;
+
+    public String getGoogleTokenURI(){
+        return getGoogleAuthUrl() + "/token";
+    }
+
+    public String getGoogleTokenInfoURI(String jwtAccessToken){
+        return UriComponentsBuilder.fromHttpUrl(getGoogleAuthUrl() + "/tokeninfo").queryParam("access_token", jwtAccessToken).toUriString();
+    }
 
 }
