+++ backend/spring-boot-3/src/main/java/back/springbootdeveloper/seungchan/repository/MemberRepository.java
@@ -9,6 +9,5 @@ import java.util.Optional;
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
-
     Optional<Member> findByEmail(String email);
 }
